refactor: split `Model` types file into multiple files
