+++ src/Model/types/AttributesAliasesMap.ts
@@ -1,0 +1,4 @@
+/**
+ * A map of attribute names to corresponding aliases, or vice versa.
+ */
+export type AttributesAliasesMap = Record<string, string>;
