+++ src/Model/types/ModelConstructorParams.ts
@@ -1,0 +1,15 @@
+import { DdbClientWrapper } from "../../DdbClientWrapper/index.js";
+import type { ModelSchemaOptions } from "../../Schema/types/index.js";
+import type { TableKeysAndIndexes, TableConstructorParams } from "../../Table/types/index.js";
+import type { Writable } from "type-fest";
+
+/**
+ * Constructor params for creating a new `Model` instance.
+ */
+export type ModelConstructorParams = Writable<
+  TableKeysAndIndexes
+    & ModelSchemaOptions
+    & Pick<TableConstructorParams, "tableName"> & {
+      ddb: DdbClientWrapper;
+    }
+>;
