+++ src/Model/types/index.ts
@@ -1,0 +1,4 @@
+export * from "./AttributesAliasesMap.js";
+export * from "./clientParamsForModel.js";
+export * from "./KeyParameters.js";
+export * from "./ModelConstructorParams.js";
